+++ src/frontend/apps/accounts/src/components/AppFooter/AppFooter.tsx
@@ -3,6 +3,7 @@
 import dynamic from 'next/dynamic';
 
 import { AppHelpMenu } from '@/components/AppHelpMenu/AppHelpMenu';
+import { AppUserMenu } from '@/components/AppUserMenu/AppUserMenu';
 
 const LaGaufre = dynamic(
   () => import('@gouvfr-lasuite/ui-kit').then((module) => module.LaGaufreV2),
@@ -11,9 +12,14 @@ const LaGaufre = dynamic(
 
 export const AppFooter = () => (
   <footer className="suite__footer">
+    <div className="suite__footer__user">
+      <AppUserMenu />
+    </div>
     <div className="suite__footer__gaufre">
-      <LaGaufre apiUrl="https://lasuite.numerique.gouv.fr/api/services"
-      widgetPath="https://static.suite.anct.gouv.fr/widgets/lagaufre.js" />
+      <LaGaufre
+        apiUrl="https://lasuite.numerique.gouv.fr/api/services"
+        widgetPath="https://static.suite.anct.gouv.fr/widgets/lagaufre.js"
+      />
     </div>
     <div className="suite__footer__help">
       <AppHelpMenu />