+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to
 ## [Unreleased]
 
 ### Added
-
+- 💄(frontend) add userMenu on footer
 - 💄(frontend) add profile home page
 - 💄(frontend) add intermediate login page
 - ✨(frontend) init the frontend app (webapp + e2e)